requirements.txt
flask==2.3.3
textblob==0.19.0
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
